Reorganize code into more modules to simplify main.rs
